style: Update some class and variable names.
