+++ LibDmd/Converter/Vni/VniFile.cs
@@ -1,130 +1,63 @@
 ﻿using System.Collections.Generic;
 using System.IO;
-using System.Linq;
+using System.Text;
 using LibDmd.Common;
-using NLog;
+using LibDmd.Input.FileSystem;
 
 namespace LibDmd.Converter.Vni
 {
-	/// <summary>
-	/// Reads the VNI file format.
-	/// 
-	/// Example here: http://vpuniverse.com/forums/files/category/84-pin2dmd-files/
-	/// Documentation here: https://github.com/sker65/go-dmd-clock/blob/master/doc/README.md
-	/// </summary>
-	public class VniFile
+	public class VniFile : AnimationSet
 	{
-		public readonly string Filename;
+		public int MaxHeight { get; protected set; }
+		public int MaxWidth { get; protected set; }
 
-		/// <summary>
-		/// File version. 1 = FSQ, 2 = VNI (but we don't really care, we fetch what we get)
-		/// </summary>
-		public readonly int Version;
-
-		public readonly VniPalette[] Palettes;
-		public readonly Dictionary<uint, Mapping> Mappings;
-		public readonly byte[][] Masks;
-		public readonly VniPalette DefaultPalette;
-		public readonly ushort DefaultPaletteIndex;
-		public readonly int NumPalettes;
-
-		private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
-
-		/// <summary>
-		/// Opens and parses the .vni file.
-		/// </summary>
-		/// <param name="filename">Path to the file</param>
 		public VniFile(string filename)
 		{
 			var fs = new FileStream(filename, FileMode.Open);
 			var reader = new BinaryReader(fs);
 
-			Mappings = null;
-			Masks = null;
-			Filename = filename;
-			Version = reader.ReadByte();
-			Logger.Trace("[vni] PAL[{1}] Read version as {0}", Version, reader.BaseStream.Position);
-
-			NumPalettes = reader.ReadUInt16BE();
-			Logger.Trace("[vni] PAL[{1}] Read number of palettes as {0}", NumPalettes, reader.BaseStream.Position);
-			Palettes = new VniPalette[NumPalettes];
-			for (var i = 0; i < NumPalettes; i++) {
-				Palettes[i] = new VniPalette(reader);
-				if (DefaultPalette == null && Palettes[i].IsDefault) {
-					DefaultPalette = Palettes[i];
-					DefaultPaletteIndex = (ushort)i;
-				}
-			}
-			if (DefaultPalette == null && Palettes.Length > 0) {
-				DefaultPalette = Palettes[0];
-			}
-
-			if (reader.BaseStream.Position == reader.BaseStream.Length) {
+			// name
+			var header = Encoding.UTF8.GetString(reader.ReadBytes(4));
+			if (header != "VPIN") {
 				reader.Close();
-				return;
+				fs.Close();
+				throw new WrongFormatException("Not a VPIN file: " + filename);
 			}
 
-			var numMappings = reader.ReadUInt16BE();
-			Logger.Trace("[vni] PAL[{1}] Read number of mappings as {0}", numMappings, reader.BaseStream.Position);
-			if (reader.BaseStream.Length - reader.BaseStream.Position < Mapping.Length * numMappings) {
-				Logger.Warn("[vni] [{1}] Missing {0} bytes for {1} masks, ignoring.", Mapping.Length * numMappings - reader.BaseStream.Length + reader.BaseStream.Position, numMappings);
-				reader.Close();
-				return;
-			}
+			// version
+			Version = reader.ReadInt16BE();
 
-			if (numMappings > 0) {
-				Mappings = new System.Collections.Generic.Dictionary<uint, Mapping>();
-				for (var i = 0; i < numMappings; i++) {
-					var mapping = new Mapping(reader);
-					Mappings.Add(mapping.Checksum, mapping);
-				}
-			} else if (numMappings == 0 || reader.BaseStream.Position == reader.BaseStream.Length) {
-				if (reader.BaseStream.Position != reader.BaseStream.Length) {
-					Logger.Warn("[vni] PAL[{1}] No mappings found but there are still {0} bytes in the file!", reader.BaseStream.Length - reader.BaseStream.Position, reader.BaseStream.Position);
-				}
-				reader.Close();
-				return;
-			}
+			// number of animations
+			var numAnimations = reader.ReadInt16BE();
 
-			var numMasks = reader.ReadByte();
-			Logger.Trace("[vni] PAL[{1}] Read number of masks as {0}", numMasks, reader.BaseStream.Position);
-			if (numMasks > 0) {
-				int maskBytes = (int)(reader.BaseStream.Length - reader.BaseStream.Position) / numMasks;
-
-				if (maskBytes != 256 && maskBytes != 512 && maskBytes != 1536) {
-					Logger.Warn("[vni] {0} bytes remaining per {1} masks.  Unknown size, ignoring.", maskBytes, numMasks);
-					reader.Close();
-					return;
-				}
-				Masks = new byte[numMasks][];
-				for (var i = 0; i < numMasks; i++) {
-					Masks[i] = reader.ReadBytesRequired(maskBytes);
-					// Logger.Trace("[{1}] Read number of {0} bytes of mask", Masks[i].Length, reader.BaseStream.Position);
+			if (Version >= 2) {
+				Logger.Trace("[vni] VNI[{1}] Skipping {0} bytes of animation indexes.", numAnimations * 4, reader.BaseStream.Position);
+				for (var i = 0; i < numAnimations; i++) {
+					reader.ReadUInt32();
 				}
 			}
 
-			if (reader.BaseStream.Position != reader.BaseStream.Length) {
-				throw new IOException("Read error, finished parsing but there are still " + (reader.BaseStream.Length - reader.BaseStream.Position) + " bytes to read.");
-			}
+			Animations = new List<Animation>(numAnimations);
+			Logger.Debug("[vni] VNI[{3}] Reading {0} animations from {1} v{2}...", numAnimations, header, Version, reader.BaseStream.Position);
 
+			MaxWidth = 0;
+			MaxHeight = 0;
+			for (var i = 0; i < numAnimations; i++) {
+				Animations.Add(new VniAnimation(reader, Version));
+				int h = Animations[i].Size.Height;
+				int w = Animations[i].Size.Width;
+				if (h > MaxHeight)
+					MaxHeight = h;
+				if (w > MaxWidth)
+					MaxWidth = w;
+			}
 			reader.Close();
-		}
-
-		public VniPalette GetPalette(uint index)
-		{
-			// TODO index bruichä
-			return Palettes.FirstOrDefault(p => p.Index == index);
-		}
-
-		public Mapping FindMapping(uint checksum)
-		{
-			Mappings.TryGetValue(checksum, out var mapping);
-			return mapping;
+			fs.Close();
 		}
 
 		public override string ToString()
 		{
-			return $"{Path.GetFileName(Filename)}: v{Version}, {Palettes.Length} palette(s), {Mappings.Count} mapping(s), {Masks.Length} mask(s)";
+			return $"VPIN v{Version}, {Animations.Count} animation(s)";
 		}
 	}
 }
